NPC attack and damage bonuses v2 - a more rules compliant version
This patch is designed to largely remove the need to duplicate loot-based weapons on NPC sheets by implementing new attack and damage bonuses based on the rules set out in AA1.

A total of three different sets of configurable attack and damage bonuses can be configured (high, low, and special).
Each type of weapon (as well as armor and shields) can then be set to use one of the three sets of bonuses.
If one of the bonuses is selected for a specific type of weapon, the matching attack bonus is added instead of the normal attribute bonus (str, dex, etc).
The matching damage bonus is added to all weapon attacks where 'apply bonus damage' is checked (this can be set for everything except grenades).

In most cases, the high and low bonuses can be set for new NPCs by following the instructions in the AA1 monster creation rules.
For example, a CR 6 melee combatant would set these values: high +16/+6, low +13/+6, melee weapons using high bonus, ranged weapons using low bonus.
You can then just drag weapons from the equipment compendium to the actor sheet and they should get the correct bonuses automatically.

A small minority of NPCs require a third set of bonuses in order to correctly set up the stat block. This is handled by including a 'special' bonus set.
The special bonus set is rarely needed, and for the majority of NPCs (especially non-humanoid aliens) it is superfluous and can be left unused.

To avoid existing NPCs from breaking, the default behaviour is for the bonuses to be set to 0 and not set against any weapon types.
If a particular bonus isn't set for a type of weapon, it falls back to the old behaviour and uses the configured attribute bonus on the item sheet.